Rename and rearrange files in lib and test directories
- All route handler tests now go in route-handlers/, since they are not
  testing code from lib/
- All key files in the handler generation flow have been moved out of
  util/ and up into lib/, since they are now the core functionality of
  the library
- All files in lib/* root have been renamed to remove verbs, exposing
  the operative methods as named exports
- WPRequest folder has been renamed to constructors/
