+++ lib/constructors/wp-request.js
