+++ lib/endpoint-factories.js
@@ -6,7 +6,7 @@
 
 var extend = require( 'node.extend' );
 var createResourceHandlerSpec = require( './resource-handler-spec' ).create;
-var makeEndpointRequest = require( './make-endpoint-request' );
+var createEndpointRequest = require( './endpoint-request' ).create;
 
 /**
  * Given an array of route definitions and a specific namespace for those routes,
@@ -25,7 +25,7 @@ function generateEndpointFactories( namespace, routeDefinitions ) {
 
 		var handlerSpec = createResourceHandlerSpec( routeDefinitions[ resource ], resource );
 
-		var EndpointRequest = makeEndpointRequest( handlerSpec, resource, namespace );
+		var EndpointRequest = createEndpointRequest( handlerSpec, resource, namespace );
 
 		// "handler" object is now fully prepared; create the factory method that
 		// will instantiate and return a handler instance
@@ -39,4 +39,6 @@ function generateEndpointFactories( namespace, routeDefinitions ) {
 	}, {} );
 }
 
-module.exports = generateEndpointFactories;
+module.exports = {
+	generate: generateEndpointFactories
+};
