+++ lib/endpoint-request.js
@@ -1,10 +1,10 @@
 'use strict';
 
 var inherit = require( 'util' ).inherits;
-var WPRequest = require( '../shared/wp-request' );
-var mixins = require( '../mixins' );
+var WPRequest = require( './constructors/wp-request' );
+var mixins = require( './mixins' );
 
-function makeEndpointRequest( handlerSpec, resource, namespace ) {
+function createEndpointRequest( handlerSpec, resource, namespace ) {
 
 	// Create the constructor function for this endpoint
 	function EndpointRequest( options ) {
@@ -51,4 +51,6 @@ function makeEndpointRequest( handlerSpec, resource, namespace ) {
 	return EndpointRequest;
 }
 
-module.exports = makeEndpointRequest;
+module.exports = {
+	create: createEndpointRequest
+};
