+++ lib/path-part-setter.js
@@ -11,7 +11,7 @@
  * @param {[type]} node [description]
  * @returns {[type]} [description]
  */
-function generatePathPartSetter( node ) {
+function createPathPartSetter( node ) {
 	// Local references to `node` properties used by returned functions
 	var nodeLevel = node.level;
 	var nodeName = node.names[ 0 ];
@@ -80,4 +80,6 @@ function generatePathPartSetter( node ) {
 	}
 }
 
-module.exports = generatePathPartSetter;
+module.exports = {
+	create: createPathPartSetter
+};
