+++ lib/resource-handler-spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var generatePathPartSetter = require( './generate-path-part-setter' );
+var createPathPartSetter = require( './path-part-setter' ).create;
 
 function addLevelOption( levelsObj, level, obj ) {
 	levelsObj[ level ] = levelsObj[ level ] || [];
@@ -20,7 +20,7 @@ function assignSetterFnForNode( handler, node ) {
 	// First level is set implicitly, no dedicated setter needed
 	if ( node.level > 0 ) {
 
-		setterFn = generatePathPartSetter( node );
+		setterFn = createPathPartSetter( node );
 
 		node.names.forEach(function( name ) {
 			// camel-case the setter name
