+++ lib/route-tree.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var namedGroupRegexp = require( './named-group-regexp' );
-var ensure = require( './ensure' );
+var namedGroupRegexp = require( './util/named-group-regexp' );
+var ensure = require( './util/ensure' );
 
 /**
  * Method to use when reducing route components array.
