+++ tests/unit/lib/constructors/wp-request.js
@@ -5,7 +5,7 @@ chai.use( require( 'sinon-chai' ) );
 var sinon = require( 'sinon' );
 var sandbox = require( 'sandboxed-module' );
 
-var WPRequest = require( '../../../../lib/shared/wp-request' );
+var WPRequest = require( '../../../../lib/constructors/wp-request' );
 var filterMixins = require( '../../../../lib/mixins/filters' );
 
 describe( 'WPRequest', function() {
@@ -405,7 +405,7 @@ describe( 'WPRequest', function() {
 
 		beforeEach(function() {
 			mockAgent = new MockAgent();
-			SandboxedRequest = sandbox.require( '../../../../lib/shared/wp-request', {
+			SandboxedRequest = sandbox.require( '../../../../lib/constructors/wp-request', {
 				requires: {
 					'superagent': mockAgent
 				}
