+++ tests/unit/route-handlers/media.js
@@ -2,7 +2,7 @@
 var expect = require( 'chai' ).expect;
 
 var WP = require( '../../../wp' );
-var WPRequest = require( '../../../lib/shared/wp-request' );
+var WPRequest = require( '../../../lib/constructors/wp-request' );
 
 describe( 'wp.media', function() {
 	var site;
